Fix teacher LSE addressing beyond 64K query rows
Promote physical query and key indices to int64 before stride products in
the CSA2 teacher LSE and shared window/compressed LSE kernels. This avoids
int32 element-offset overflow for flattened H64/D512 query storage.

Add regressions covering dense and candidate teacher LSE, SBHD and THD,
and query rows on both sides of the signed int32 offset boundary.

Validation: 4 GPU regressions passed on a DLCluster H100 NVL using a real
65538-row H64/D512 allocation.

+++ megatron/core/transformer/experimental_attention_variant/csa_utils/csa2_indexer.py
@@ -304,7 +304,8 @@ if HAVE_TRITON:
         TILE_H: tl.constexpr,
         TILE_K: tl.constexpr,
     ):
-        row = tl.program_id(0)
+        # This launch spans the full flattened batch, before loss query chunking.
+        row = tl.program_id(0).to(tl.int64)
         h = tl.program_id(1) * TILE_H + tl.arange(0, TILE_H)
         d = tl.arange(0, DIM)
         qv = tl.load(q + (row * HEADS + h[:, None]) * DIM + d[None, :], h[:, None] < HEADS, other=0)
@@ -315,7 +316,7 @@ if HAVE_TRITON:
             keys, valid = _key_ids(
                 ids, counts, starts, visible, row, cols, WIDTH, ID_WIDTH, MODE, BLOCK_SIZE
             )
-            kv = tl.load(k + keys[None, :] * DIM + d[:, None], valid[None, :], other=0)
+            kv = tl.load(k + keys[None, :].to(tl.int64) * DIM + d[:, None], valid[None, :], other=0)
             logits = tl.where(valid[None, :], tl.dot(qv, kv) * SCALE, -float("inf"))
             new_max = tl.maximum(maximum, tl.max(logits, 1))
             old_scale = tl.where(maximum > -float("inf"), tl.exp(maximum - new_max), 0.0)
